Move `WelcomeTheme` out of init package
